store_index.py
from dotenv import load_dotenv
import os
from pinecone import Pinecone
from src.helper import pdf_loader, filter_to_minimal_docs, text_split, download_embeddings
from pinecone import ServerlessSpec
from langchain_pinecone import PineconeVectorStore
load_dotenv()

PINECONE_API_KEY = os.getenv("PINECONE_API_KEY")
GROQ_API_KEY = os.getenv("GROQ_API_KEY")

os.environ["PINECONE_API_KEY"] = PINECONE_API_KEY
os.environ["GROQ_API_KEY"] = GROQ_API_KEY

extracted_data = pdf_loader(data="data/")
filter_data = filter_to_minimal_docs(extracted_data)
text_chunks = text_split(filter_data)

embeddings = download_embeddings()

Pinecone_api_key = PINECONE_API_KEY
pc = Pinecone(api_key=Pinecone_api_key)

index_name = "ai-chatbot"

if index_name not in pc.list_indexes().names():
    pc.create_index(
        name=index_name,
        dimension=384,
        metric="cosine",
        spec=ServerlessSpec(cloud="aws", region="us-east-1"),
    )

index = pc.Index(index_name)

docsearch = PineconeVectorStore.from_documents(
    documents=text_chunks,
    embedding=embeddings,
    index_name=index_name,
    )
